+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Table chart component
 - Map chart component
 
+### Changed
+- Improve filters layout
+
 ## [0.3.0] - 2023-03-14
 ### Added
 - Select filter type with static and dynamic values